fix: fix tests and request.rest
